+++ pgmax/factors/logical.py
@@ -1,13 +1,16 @@
 """Defines a logical factor"""
 
-from dataclasses import dataclass
-from typing import Mapping, Sequence, Union
+import functools
+from dataclasses import dataclass, field
+from typing import Mapping, Optional, Sequence, Union
 
 import jax
 import jax.numpy as jnp
 import numpy as np
+from jax.nn import log_sigmoid, sigmoid
 
 from pgmax import utils
+from pgmax.bp import bp_utils
 from pgmax.fg import nodes
 
 
@@ -37,24 +40,26 @@ class LogicalWiring(nodes.Wiring):
     children_edge_states: Union[np.ndarray, jnp.ndarray]
 
     def __post_init__(self):
-        logical_factor_indices = self.parents_edge_states[:, 0]
-        num_logical_factors = self.children_edge_states.shape[0]
+        if self.children_edge_states.shape[0] > 0:
+            logical_factor_indices = self.parents_edge_states[:, 0]
+            num_logical_factors = self.children_edge_states.shape[0]
 
-        if logical_factor_indices.max() >= num_logical_factors:
-            raise ValueError(
-                f"The highest OR factor index must be {num_logical_factors - 1}"
-            )
+            if logical_factor_indices.max() >= num_logical_factors:
+                raise ValueError(
+                    f"The highest OR factor index must be {num_logical_factors - 1}"
+                )
 
-        if np.unique(logical_factor_indices).shape[0] != num_logical_factors:
-            raise ValueError(
-                f"There must be {num_logical_factors} different OR factor indices"
-            )
+            if np.unique(logical_factor_indices).shape[0] != num_logical_factors:
+                raise ValueError(
+                    f"There must be {num_logical_factors} different OR factor indices"
+                )
 
-        inference_arguments = {
+    @property
+    def inference_arguments(self) -> Mapping[str, np.ndarray]:
+        return {
             "parents_edge_states": self.parents_edge_states,
             "children_edge_states": self.children_edge_states,
         }
-        object.__setattr__(self, "inference_arguments", inference_arguments)
 
 
 @dataclass(frozen=True, eq=False)
@@ -67,6 +72,8 @@ class LogicalFactor(nodes.Factor):
             (1) The variables are not all binary
             (2) There are less than 2 variables
     """
+
+    log_potentials: np.ndarray = field(init=False, default=np.empty((0,)))
 
     def __post_init__(self):
         if not np.all([variable.num_states == 2 for variable in self.variables]):
@@ -144,7 +151,12 @@ def concatenate_logical_wirings(
         ValueError: if the list of LogicalWirings is empty
     """
     if len(logical_wirings) == 0:
-        raise ValueError("The list of LogicalWirings is empty")
+        return LogicalWiring(
+            edges_num_states=np.empty((0,), dtype=int),
+            var_states_for_edges=np.empty((0,), dtype=int),
+            parents_edge_states=np.empty((0, 2), dtype=int),
+            children_edge_states=np.empty((0,), dtype=int),
+        )
 
     # Note: this correspomds to all the factor_to_msgs_starts for the EnumerationFactors
     num_edge_states_cumsum = np.insert(
@@ -177,3 +189,129 @@ def concatenate_logical_wirings(
 @dataclass(frozen=True, eq=False)
 class ORFactor(LogicalFactor):
     pass
+
+
+@functools.partial(jax.jit, static_argnames=("temperature"))
+def pass_OR_fac_to_var_messages(
+    vtof_msgs: jnp.ndarray,
+    parents_edge_states: jnp.ndarray,
+    children_edge_states: jnp.ndarray,
+    temperature: float,
+    log_potentials: Optional[jnp.ndarray] = None,
+) -> jnp.ndarray:
+
+    """Passes messages from ORFactors to Variables.
+
+    Args:
+        vtof_msgs: Array of shape (num_edge_state,). This holds all the flattened variable to all the ORFactors messages,
+        parents_edge_states: Array of shape (num_parents, 2)
+            parents_edge_states[ii, 0] contains the global ORFactor index,
+            parents_edge_states[ii, 1] contains the message index of the parent variable's state 0.
+            Both indices only take into account the ORFactors of the FactorGraph
+            The parent variable's state 1 is parents_edge_states[ii, 2] + 1
+        children_edge_states: Array of shape (num_factors,)
+            children_edge_states[ii] contains the message index of the child variable's state 0
+            The child variable's state 1 is children_edge_states[ii, 1] + 1
+        temperature: Temperature for loopy belief propagation.
+            1.0 corresponds to sum-product, 0.0 corresponds to max-product.
+
+    Returns:
+        Array of shape (num_edge_state,). This holds all the flattened ORFactors to variable messages.
+    """
+    num_factors = children_edge_states.shape[0]
+
+    factor_indices = parents_edge_states[..., 0]
+
+    parents_tof_msgs = (
+        vtof_msgs[parents_edge_states[..., 1] + 1]
+        - vtof_msgs[parents_edge_states[..., 1]]
+    )
+    children_tof_msgs = (
+        vtof_msgs[children_edge_states + 1] - vtof_msgs[children_edge_states]
+    )
+
+    # We treat the max-product case separately.
+    if temperature == 0.0:
+        # Get the first and second argmaxes for the incoming parents messages of each factor
+        _, first_parents_argmaxes = bp_utils.get_maxes_and_argmaxes(
+            parents_tof_msgs, factor_indices, num_factors
+        )
+        _, second_parents_argmaxes = bp_utils.get_maxes_and_argmaxes(
+            parents_tof_msgs.at[first_parents_argmaxes].set(bp_utils.NEG_INF),
+            factor_indices,
+            num_factors,
+        )
+
+        parents_tof_msgs_pos = jnp.maximum(0.0, parents_tof_msgs)
+        sum_parents_tof_msgs_pos = (
+            jnp.full(shape=(num_factors,), fill_value=0.0)
+            .at[factor_indices]
+            .add(parents_tof_msgs_pos)
+        )
+
+        # Outgoing messages to parents variables
+        # See https://arxiv.org/pdf/2111.02458.pdf, Appendix C.3
+        parents_msgs = jnp.minimum(
+            children_tof_msgs[factor_indices]
+            + sum_parents_tof_msgs_pos[factor_indices]
+            - parents_tof_msgs_pos,
+            jnp.maximum(0.0, -parents_tof_msgs[first_parents_argmaxes][factor_indices]),
+        )
+        parents_msgs = parents_msgs.at[first_parents_argmaxes].set(
+            jnp.minimum(
+                children_tof_msgs
+                + sum_parents_tof_msgs_pos
+                - parents_tof_msgs_pos[first_parents_argmaxes],
+                jnp.maximum(0.0, -parents_tof_msgs[second_parents_argmaxes]),
+            )
+        )
+
+        # Outgoing messages to children variables
+        children_msgs = sum_parents_tof_msgs_pos + jnp.minimum(
+            0.0, parents_tof_msgs[first_parents_argmaxes]
+        )
+    else:
+
+        def g(x):
+            # assert jnp.all(x >= 0)
+            return jnp.where(
+                x == 0.0,
+                0.0,
+                x + temperature * jnp.log(1.0 - jnp.exp(-x / temperature)),
+            )
+
+        log_sig_parents_tof_msgs = -temperature * log_sigmoid(
+            -parents_tof_msgs / temperature
+        )
+        sum_log_sig_parents_tof_msgs = (
+            jnp.full(shape=(num_factors,), fill_value=0.0)
+            .at[factor_indices]
+            .add(log_sig_parents_tof_msgs)
+        )
+        g_sum_log_sig_parents_minus_id = g(
+            sum_log_sig_parents_tof_msgs[factor_indices] - log_sig_parents_tof_msgs
+        )
+
+        # Outgoing messages to parents variables
+        parents_msgs = -temperature * jnp.log(
+            sigmoid(g_sum_log_sig_parents_minus_id / temperature)
+            + sigmoid(-g_sum_log_sig_parents_minus_id / temperature)
+            * jnp.exp(-children_tof_msgs[factor_indices] / temperature)
+        )
+
+        # Outgoing messages to children variables
+        children_msgs = g(sum_log_sig_parents_tof_msgs)
+
+    # Special case: factors with a single parent
+    num_parents = jnp.bincount(factor_indices, length=num_factors)
+    first_elements = jnp.concatenate(
+        [jnp.zeros(1, dtype=int), jnp.cumsum(num_parents)]
+    )[:-1]
+    parents_msgs = parents_msgs.at[first_elements].set(
+        jnp.where(num_parents == 1, children_tof_msgs, parents_msgs[first_elements]),
+    )
+
+    ftov_msgs = jnp.zeros_like(vtof_msgs)
+    ftov_msgs = ftov_msgs.at[parents_edge_states[..., 1] + 1].set(parents_msgs)
+    ftov_msgs = ftov_msgs.at[children_edge_states + 1].set(children_msgs)
+    return ftov_msgs
